`*.after_request` hook: always pass `$info`
Callbacks hooked into the `[curl|fsockopen].after_request` hook would previously only receive the `$info` parameter for blocking requests. For non-blocking requests, the parameter would be omitted.

This made creating callbacks a little awkward, especially with both parameters being passed by reference.

This PR changes this to:
* Always pass the `$info` parameter by reference. When the request is non-blocking, this will just be a plain empty array.
* Update the hook documentation to reflect this change.

Note: this _could_ be considered a breaking change, so discussion may be warranted on whether this change can go into a minor or needs to wait for a major release.

Fixes 321